feat/admin-users: added a route to get all users in the admin panel, BUG-FIX: UPDATED THE DETAILS PAGE DUE TO INFINITE LOOP OF RE-RENDERING
