+++ client/src/pages/admin/articles/landing/index.js
@@ -33,7 +33,6 @@ const ManageArticles = () => {
                         pathname: `/admin/article/edit/${article._id}`
                     }} className={styles.edit}>Edit</Link></td>
                     <td><Link to="#" onClick={() => deleteArticle(article._id)} className={styles.delete}>Delete</Link></td>
-                    <td><Link to="#" className={styles.publish}>Publish</Link></td>
                 </tr>
             )
         })
