+++ server/routes/article.js
@@ -4,13 +4,11 @@ const auth = require('../utils/auth');
 
 router.get('/all', controllers.article.get.all);
 
-// router.get('/create', controllers.article.get.create);
 router.post('/create', auth(), controllers.article.post.create);
 
 router.get('/details/:id', controllers.article.get.details);
 
 router.get('/edit/:id', controllers.article.get.edit);
-// router.post('/edit/:id', controllers.article.post.edit);
 router.put('/edit/:id', controllers.article.post.edit)
 
 router.get('/delete/:id', controllers.article.get.delete);
