+++ server/routes/user.js
@@ -1,11 +1,9 @@
 const router = require('express').Router();
 const controllers = require('../controllers');
-// const auth = require('../utils/auth');
+
+router.get('/all', controllers.user.get.all);
 
 router.get('/profile', controllers.user.get.profile);
-
-// router.get('/login', controllers.user.get.login);
-// router.get('/register', controllers.user.get.register);
 
 router.get('/logout', controllers.user.get.logout);
